Disable global modes upon entry into the minibuffer
This is one way to ensure that mode keymaps that are promoted to
terminal local (highest priority) keymaps play nicely with the
minibuffer.

This approach does not work for local modes since they are not active
in the minibuffer to begin with, so we cannot exit them and thereby
demote the keymap (which retains priority in the minibuffer even
though the minor mode is not enabled there!).