Side-channel mechanism to de-prioritize mode keymaps
Lithium modes build on top of Emacs's existing minor mode
infrastructure and use their control protocols to enable and disable
keymaps in buffers. These mode keymaps also need to have higher
priority than minor mode keymaps, however. The appropriate means to do
that for our purposes is to promote the keymaps to "overriding
terminal mode maps" which means they take priority over all other
keymaps. But this set of overriding keymaps is checked outside of any
minor mode controls, and thus causes these keymaps to be in effect and
to take precedence over other keymaps even in buffers where the minor
mode is not enabled, if we happen to be within the mode elsewhere.

We introduce a mechanism to suspend and reinstate mode keymaps from
the list of overriding terminal mode maps, which can be availed of via
hooks. The main need for this at the moment is in the minibuffer,
since neither global nor local lithium modes are active there, yet the
minibuffer may be entered while these modes are active, in which case
those keymaps would still be in effect in the minibuffer.