Handle keymap promotion in wrapping mode-defining macros
The wrapping macros are the right place to handle entry and exit from
lithium modes, since these are called just once in each instance,
whether for local modes or global modes.

Formerly, we were handling keymap promotion in the underlying minor
mode macro since the keymap is defined there and it seemed convenient.
But due to the dichotomy of local vs global modes, both of which
employ the same underlying minor mode definition macro, but which have
different semantics for entry and exit, it is not possible to do it
there in a way that is simultaneously well-defined in both of these
cases. So, we bring this common logic up to the appropriate places in
the distinct surface macros.