Buffer-local state and mode inference
Formerly we were relying on suspending and reinstating overriding
keymaps when entering and leaving the minibuffer, to ensure that the
overriding map is not loaded while there. This was inadequate as we
also need to "suspend" the map when visiting other buffers where the
mode is not active, or where, perhaps, a different mode is active.

Instead of relying on a global state variable tracking the suspended
keymap, which is brittle, we now rely on buffer change hooks to always
promote the appropriate keymap upon visiting a buffer. This is more
robust as the state is always locally determined, inferred from global
and local modes and whether we are in the minibuffer, without needing
any separate state aside from the modal configuration.