Maintain mode metadata instead of attempting to infer keymap name
Dynamically constructing the expected name of a variable is brittle in
general, and in this case it doesn't always work since local modes and
global modes both use the keymap defined by the local minor mode,
which differs from the name of the mode in the global case. This was
causing an error resembling `....-mode-map not defined`, as the map in
this case is named `...-local-mode-map`.

We now instead construct an explicit metadata struct containing the
name of the mode as well as the relevant keymap, so that we can
consult this struct and promote or demote the keymap explicitly.