Use a stack of modes
Instead of having a distinct "current" global and local mode, and the
need to determine the appropriate overriding map by considering these,
we now use a mode stack. The top of the stack is effectively the
"current" mode as its keymap is promoted as the only overidding map.
When popped, the next mode below it, if any, would have its keymap
promoted. This more robustly models the expected behavior of global vs
local modes, and the ability for even local modes to override other
local modes without the prior modes needing to be escaped first.