Don't compete with other overriding maps
Lithium is meant to be used persistently but not "at rest," in the
sense that although we use these modes for prolonged periods, they are
still specialized rather than general --- Evil Normal state is an
example of a general state which we would use "at rest" in this sense.
Vanilla Emacs is another such general, at-rest state. Lithium uses the
overriding terminal map as it generally needs to override at-rest
keybindings, even in cases where they assume high priority like in
Magit buffers. But at the same time, it must be overridden by other
in-motion keymaps that seek to be overriding (like Hydra and
Transient), as these typically are not persistent, and we prefer them
to take priority on that basis. Something like:

  at-rest < persistent-in-motion < ephemeral-in-motion