Ignore "mundane" errors
We want Lithium to get out of the way (i.e., exit) in case of an
actual error, but in Emacs, some ordinary events, like attempting to
scroll past the top of the buffer, are signaled as errors. If these
errors occur, we want to ignore them and persist in the Lithium mode.

Incidentally, Emacs doesn't really treat these as errors, either, as
it doesn't enter the debugger on such errors if
`toggle-debug-on-error` is enabled. It would be nice if Emacs afforded
a standard way to identify such "mundane" errors, but as I wasn't able
to find something like this, we explicitly enumerate the ones we know
of, as there don't seem to be all that many.