fix bug with overriding keymap interaction with edebug
Entering an edebug session while a lithium mode is active was in some
cases leaving the local accounting of the overriding map out of sync
with Emacs's actual overriding map, so that we no longer control the
overriding map though it is active. As a fix, if there is an
overriding map present that is locally unaccounted for, we now check
to see if it's a local map and bring it back into the accounting.