also use window-configuration-change-hook to re-eval overriding map
We already use `window-buffer-change-functions` and
`window-selection-change-functions` to ensure that newly selected
buffers evaluate the overriding map and demote/promote the appropriate
one. But in the case where a new buffer is implicitly created, e.g.
`cider-jack-in` when using Clojure, it was evading these hooks. With
the addition of `window-configuration-change-hook`, it now catches all
known cases.

The other option is to use `buffer-list-update-hook` instead of these
various hooks in tandem, as that is likely to be comprehensive. But if
we did that, we'd either need to exhaustively call eval map in every
buffer in that list each time it changes, or cache the list each time
and invoke it on the diff. The latter seems viable, but it entails
state and seems undesirable from that perspective. Being able to do it
locally without maintaining state (approximating a "new buffer" hook,
which doesn't seem to exist) seems better.