Simplify macro delegation and produced expansion
The expansion produced by the lithium mode definition formerly had
complicated keybindings in the resulting minor mode map. This was
making it hard to debug, as the keymaps were inscrutable. It was also
making it difficult to override variables like `this-command` (to give
Emacs the impression that the original command indicated in the mode
specification is the one being executed, which should allow lithium's
operation to be transparent, as this information is used in, e.g.,
`evil-repeat`).

This simplifies the produced expansion to just the actual lambda that
will be executed when a key is pressed — that is, the one that calls
pre-hooks, runs the command with error handling, and then calls
post-hooks. The keymap is still not as easy to read as an ordinary
minor mode map or even a Hydra keymap, but that is unavoidable as we
do in fact do more error handling and calling of hooks in addition to
calling the specified function, and it seems correct not to try and
obscure that.

To accomplish this, it was necessary to rewrite some of the `defun`s
that were being used as `defmacro`s instead.