Propagate `this-command` to the wrapping lithium lambda
This should make Lithium's operation more transparent, by preserving
the impression, for the rest of Emacs, that the command being executed
is the one indicated in the mode specification (rather than the
wrapping Lithium lambda which runs the command, does error handling,
and calls hooks).

In particular, this variable is relied upon in `evil-repeat` as part
of determining whether a command is repeatable, as used in Symex.