change the status badge from melpazoid to the test job for now
The melpazoid check is more thorough, but it currently flags an issue
concerning the use of globalized modes that does not appear to be a
legitimate issue. The globalized mode is used as part of the
implementation rather than as something that should be invoked
directly by the user, so the suggestion to add an autoload declaration
is probably not applicable.

The link shows the full CI results, though, so we can still see the
melpazoid results there.