move pre-exit hook to after the action and right before exit
This was causing an issue where the mode was exiting before the
command was executed, which led to some assumptions in the action not
holding. For instance, the `c`hange command in Symex was losing the
selected symex and so would select afresh and change the topmost at
point instead of the originally selected one.

Another option could be to move the symex mode exit to the post-exit
hook, but it seems more correct, anyway, for the pre-exit hook to run
_after_ the action.